updated blog routes + removed card demo
