+++ website/src/css/blog_card.css
